refactor(modal): create focusTrap composable and implement in Modal
