+++ components/compositions/events.ts
@@ -1,4 +1,5 @@
-import { onMounted, onUnmounted } from 'vue'
+import { nextTick, onMounted, onUnmounted, Ref, ref } from 'vue'
+import { createFocusTrap, FocusTrap, Options } from 'focus-trap'
 
 export function handleEsc(func: Function) {
   function handleKeydown(event: KeyboardEvent) {
@@ -8,4 +9,24 @@ export function handleEsc(func: Function) {
   }
   onMounted(() => document.addEventListener('keydown', handleKeydown))
   onUnmounted(() => document.removeEventListener('keydown', handleKeydown))
+}
+
+export function handleFocusTrap(el: Ref<HTMLElement | null>) {
+  const trap: Ref<FocusTrap | null> = ref(null)
+  const options: Options = {
+    allowOutsideClick: true,
+    escapeDeactivates: false,
+  }
+
+  onMounted(() => {
+    // next tick to handle any conditionally rendered, focusable elements
+    nextTick(() => {
+      if (!el.value) {
+        return
+      }
+      trap.value = createFocusTrap(el.value, options)
+      trap.value.activate()
+    })
+  })
+  onUnmounted(() => trap.value?.deactivate())
 }
