+++ components/ui/Modal/Modal.vue
@@ -1,9 +1,8 @@
 <template src="./Modal.html"></template>
 
 <script lang="ts">
-import Vue, { nextTick, onMounted, onUnmounted, ref, Ref } from 'vue'
-import { createFocusTrap, FocusTrap, Options } from 'focus-trap'
-import { handleEsc } from '~/components/compositions/events'
+import Vue, { ref, Ref } from 'vue'
+import { handleEsc, handleFocusTrap } from '~/components/compositions/events'
 
 export default Vue.extend({
   props: {
@@ -26,29 +25,14 @@ export default Vue.extend({
     },
   },
   setup(props, { emit }) {
-    const trap: Ref<FocusTrap | null> = ref(null)
     const modal: Ref<HTMLElement | null> = ref(null)
 
     function close() {
       emit('close')
     }
 
-    onMounted(() => {
-      const options: Options = {
-        allowOutsideClick: true,
-        escapeDeactivates: false,
-      }
-
-      nextTick(() => {
-        if (!modal.value) {
-          return
-        }
-        trap.value = createFocusTrap(modal.value, options)
-        trap.value.activate()
-      })
-    })
-    onUnmounted(() => trap.value?.deactivate())
     handleEsc(close)
+    handleFocusTrap(modal)
 
     return {
       modal,
